refactor(architecture): move codebase scan into engine
