+++ src/core/code_audit/test_topology.rs
@@ -9,7 +9,7 @@ use std::path::Path;
 use super::conventions::AuditFinding;
 use super::findings::{Finding, Severity};
 use crate::extension::{self, ExtensionManifest};
-use crate::utils::codebase_scan::{self, ExtensionFilter, ScanConfig};
+use crate::engine::codebase_scan::{self, ExtensionFilter, ScanConfig};
 
 #[derive(Debug, Clone, serde::Deserialize, Default)]
 pub struct AuditRulesConfig {
