+++ src/core/engine/codebase_scan.rs
