+++ src/core/engine/mod.rs
@@ -8,6 +8,7 @@
 //! to build their orchestration logic.
 
 pub mod baseline;
+pub mod codebase_scan;
 pub mod executor;
 pub mod pipeline;
 pub mod symbol_graph;
